fix: move images to data.js
