+++ src/components/atoms/images/ContactImage.js
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { theme } from "../../../data/theme";
-import contactimg from "../../../styles/images/contactimg.jpg";
+import contactimg from "../../../data/images/contactimg.jpg";
 
 const StyledContactImage = styled.img`
   height: 80%;
